// types/character.d.ts
export type CharacterState = 'idle' | 'walking' | 'crouching' | 'jumping' | 'attacking' | 'blocking' | 'hitstun' | 'blockstun' | 'knocked_down' | 'special_move' | 'super_move' | 'parrying';
export interface CharacterStateData {
    priority: number;
    cancellable: boolean;
}
export type CharacterStates = Record<CharacterState, CharacterStateData>;